tamilhackdjango/requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.1.1
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
numpy==1.23.4
opencv-python==4.6.0.66
Pillow==9.3.0
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
wikipedia==1.4.0
